Support feature enable-dynamic-library
